Move allocator_state.hpp from public API to impl
Resolves #37232. The AllocatorState class was an internal implementation
detail exposed in the public tt-metalium API surface solely to support
unit_mesh_utils in TTNN.

- Move allocator_state.hpp from tt_metal/api/tt-metalium/ to
  tt_metal/impl/allocator/
- Add Allocator::synchronize_state_from() convenience method that
  encapsulates the extract+merge+override pattern
- Refactor unit_mesh_utils.cpp to use the new method, removing its
  dependency on allocator_state.hpp
- Update all internal includes to use relative paths
