+++ tt_metal/impl/allocator/allocator_state.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <tt-metalium/allocator_state.hpp>
+#include "allocator_state.hpp"
 #include <tt-metalium/allocator.hpp>
 #include <tt_stl/assert.hpp>
 #include <enchantum/enchantum.hpp>
