+++ ttnn/core/tensor/unit_mesh/unit_mesh_utils.cpp
@@ -8,7 +8,6 @@
 #include "ttnn/distributed/tensor_topology.hpp"
 #include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/mesh_buffer.hpp>
-#include <tt-metalium/allocator_state.hpp>
 #include <tt-metalium/allocator.hpp>
 #include <tt_stl/assert.hpp>
 
@@ -20,14 +19,15 @@ void synchronize_parent_allocator_with_submeshes(tt::tt_metal::distributed::Mesh
     auto* parent_allocator = parent_mesh->allocator().get();
     TT_FATAL(parent_allocator != nullptr, "Parent mesh must have an allocator");
 
-    tt::tt_metal::AllocatorState merged_state;
+    std::vector<tt::tt_metal::Allocator*> submesh_allocators;
+    submesh_allocators.reserve(parent_mesh->get_submeshes().size());
     for (const auto& submesh : parent_mesh->get_submeshes()) {
         auto* submesh_allocator = submesh->allocator().get();
         TT_FATAL(submesh_allocator != nullptr, "Submesh must have an allocator");
-        merged_state.merge(submesh_allocator->extract_state());
+        submesh_allocators.push_back(submesh_allocator);
     }
 
-    parent_allocator->override_state(merged_state);
+    parent_allocator->synchronize_state_from(submesh_allocators);
 }
 
 }  // namespace
